Use floats for complex numbers and expand array

+++ src/complex_g3.c
@@ -2,25 +2,26 @@
 * \brief Functions for complex_g3.h
 */
 
-#include "decoder.h"
+#include "complex_g3.h"
 
-complex_g3_t complex_from_polar(double r, double theta_radians)
+complex_g3_t complex_from_polar(float r, float theta_radians)
 {
     complex_g3_t result;
 
-    result.re = r * cos(theta_radians);
-    result.im = r * sin(theta_radians);
+    result.re = r * cosf(theta_radians);
+    result.im = r * sinf(theta_radians);
 
     return result;
 }
 
-double complex_magnitude(complex_g3_t c)
+float complex_magnitude(complex_g3_t c)
 {
-    return sqrt(c.re*c.re + c.im*c.im);
+    return sqrtf(c.re * c.re + c.im * c.im);
 }
 
-double complex_decibels(complex_g3_t c){
-    return 20 * log10(complex_magnitude(c));
+float complex_decibels(complex_g3_t c)
+{
+    return 20.0f * log10f(complex_magnitude(c));
 }
 
 complex_g3_t complex_add(complex_g3_t left, complex_g3_t right)
@@ -47,8 +48,8 @@ complex_g3_t complex_mult(complex_g3_t left, complex_g3_t right)
 {
     complex_g3_t result;
 
-    result.re = left.re*right.re - left.im*right.im;
-    result.im = left.re*right.im + left.im*right.re;
+    result.re = left.re * right.re - left.im * right.im;
+    result.im = left.re * right.im + left.im * right.re;
 
     return result;
 }
