+++ src/complex_g3.h
@@ -10,7 +10,6 @@
 #include <math.h>
 
 /* Our Headers */
-#include "complex_g3.h"
 #include "global_parameters.h"
 
 /*! \typedef complex_g3_t
@@ -18,34 +17,33 @@
 */
 typedef struct complex_g3_t
 {
-	double re;	/*!< Real component of a complex number */
-	double im;	/*!< Imaginary component of a complex number */
+        float re;       /*!< Real component of a complex number */
+        float im;       /*!< Imaginary component of a complex number */
 } complex_g3_t;
 
 /* Function Definitions */
 
-/*! \fn complex_g3_t complex_from_polar (double r, double theta_radians)
+/*! \fn complex_g3_t complex_from_polar (float r, float theta_radians)
 * \param r Magnitude of the input polar coordinate
 * \param theta_radians Angle/Phase of the input polar coordinate
 * \returns A complex number using the input information
 * \brief Converts to a \p complex_g3_t from a polar coordinate
 */
-complex_g3_t complex_from_polar(double r, double theta_radians);
+complex_g3_t complex_from_polar(float r, float theta_radians);
 
-/*! \fn double complex_magnitude (complex_g3_t c)
+/*! \fn float complex_magnitude (complex_g3_t c)
 * \param c Complex number
 * \returns The magnitude of the input complex number
 * \brief Computes the magnitude of a complex number
 */
-double complex_magnitude(complex_g3_t c);
+float complex_magnitude(complex_g3_t c);
 
-/*! \fn double complex_decibels (complex_g3_t c)
+/*! \fn float complex_decibels (complex_g3_t c)
 * \param c Complex number
 * \returns The decibels of the input complex number
 * \brief Computes the decibels of a complex number
 */
-double complex_decibels(complex_g3_t c);
-
+float complex_decibels(complex_g3_t c);
 
 /*! \fn complex_g3_t complex_add (complex_g3_t left, complex_g3_t right)
 * \param left LHS complex number
